Fix: Correct all import paths after component reorganization
- Fix import paths in PersonalInfoStep.tsx: Remove unused TicketingFundingStep reference
- Fix import paths in BusinessFinancialStep.tsx: Update relative paths for new folder structure
- Fix import paths in AllUploadsStep.tsx: Update relative paths for new folder structure
- Fix import paths in SummaryStep.tsx: Update relative paths for new folder structure
- Remove unused TicketingFundingStep component from PersonalInfoStep JSX
- All TypeScript compilation errors resolved
- No linter errors remaining
