+++ src/components/step1/PersonalInfoStep.tsx
@@ -6,7 +6,6 @@ import TextField from '../customComponents/TextField';
 import StepTitle from '../customComponents/StepTitle';
 import { useValidation } from '../../contexts/ValidationContext';
 import CompanyInfoStep from './CompanyInfoStep';
-import TicketingFundingStep from '../TicketingFundingStep';
 
 const PersonalInfoStep: React.FC = () => {
   const dispatch = useDispatch();
@@ -127,7 +126,6 @@ const PersonalInfoStep: React.FC = () => {
 
       <CompanyInfoStep />
       
-      <TicketingFundingStep />
       <p className="text-sm text-gray-500 my-4 text-center ">By filling this form, you agree to SoundCheck Capital <a href="https://soundcheckcapital.com/terms-of-service" target="_blank" rel="noopener noreferrer" className="text-blue-500">Terms of Service</a> and <a href="https://soundcheckcapital.com/privacy-policy" target="_blank" rel="noopener noreferrer" className="text-blue-500">Privacy Policy</a></p>
